+++ CHANGELOG.md
@@ -9,6 +9,12 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v5.16.2
+
+### Fixed
+
+- Fix using date validation with other fields as comparison
+
 ## v5.16.1
 
 ### Fixed